cambiar doubles tiemer

+++ timer.h
@@ -5,13 +5,13 @@
 
 void timer1_temporizador_iniciar(void);
 void timer1_temporizador_empezar(void);
-double timer1_temporizador_leer(void);
-double timer1_temporizador_parar(void);
+unsigned int timer1_temporizador_leer(void);
+unsigned int timer1_temporizador_parar(void);
 
 void timer0_temporizador_iniciar(void);
 void timer0_temporizador_empezar(void);
-double timer0_temporizador_leer(void);
-double timer0_temporizador_parar(void);
+unsigned int timer0_temporizador_leer(void);
+unsigned int timer0_temporizador_parar(void);
 void timer0_temporizador_periodico(unsigned int periodo);
